Fix lost incremented proposal on nack tie
When receiving nack tie and inrementing promise there is a period
where the incremented promise does not match the highest proposed
decree. We do this to avoid contention of competing decrees. This
can lead to lost decrees where proposer forgets the actual highest
decree it proposed. We solve this by having the acceptor return a
promise response by hijacking any lower recieved decree with the
highest promised decree iff the author of the two decrees is the
same.