Fix accept decree sent to a new replica
Issue is that a minimum accept must be configured on new replica else the
replica will be at risk of accepting a lower than allowed decree. However we
must also ensure that accepted decree has empty content else it may be
considered a stale accept decree and resend the accept in a future round.