Allow safe repurposing of replica
Issue is that if a replica is already a member of a different replica set when
it is being added as a new member of a separate replica set, then care must be
taken to ensure in-memory and on-disk state are consistent. This means promised
and accepted decrees must correspond to values in the current replica set.

This commit achieves this by first isolating the new replica, then sending over
replicated files, and finally reopening the replica to replicas in the replica
set.

Fixes #3