Fix prepare to reply tie if root decree is equal
Issue is that if do not consider equal root decrees as a tie on prepare then it
is possible to enter a scenario where a set of replicas cannot contribute
progress. As an example, let's assume we have 5 replicas {A,B,C,D,E} and {A,B}
contend with each other to drive up the decree number. If both {A,B} crash,
then {C,D,E} should still be able to propose votes since they have enough
numbers can form quorum. Thus we must allow {C,D,E} to enter the contention
late.