Fix missing decree scenario
Issue is that there exists another case where it is possible to skip over a
decree. This can happen with two contending decrees, lets say decree A and
decree B which are both proposed from 2 different replicas. Let's say A's
proposer gets a quorum of promises. Now B's proposer is still waiting on
promises, but because the proposer currently doesn't distinguish between fresh
promises and dirty promises it doesn't realize that A is actually the current
decree with quorum of promises. Then B's proposer would release the B decree
when it sends out an accept which will never reach quorum in the second phase
of commit. Thus a decree is lost.

Fixes #11