Fixes resource constraints on queue rollover
Issue is that random read/write access into large files is significantly slower
than a single append operation. This manifested itself as degraded performance
when executing many decrees in short succession. This change switches from a
circular queue based on a single file to a bufferd queue based on dual files.
The motivation is that it removes the necessity for random file accesses and
thus improves performance during high contention workloads.

Fixes #13