Fix possible hang during remove replica
Issue is that there was an assumption that a decree's handler is
executed on the same decree that proposed the remove. This is not
necessarily the case if the decree's author is not propogated when
flushing a wedged decree. Alternative solution is to ensure that author
is properly forwarded in all scenarios.

Fixes #14