Fix handler to signal iff it was proposer
Issue is that if handler sends a signal regardless of whether it was the
proposer then there exists a window when mutiple proposers are waiting
for different signals and then one handler sends a signal when it should
not.  Instead if we couple the author with the decree then we can
correctly signal only when we are the author of the decree.

Fixes #14