Fix race condition causing unaccepted decree
Issue is that there is a window where a proposer may propose a decree n.
During this time another decree n is written into the ledger and third
decree n+1 is proposed. Finally the acceptor receives the request for
decree n which it must reject because the decree is lower than the
highest promised decree n+1. However it may be the case that the
proposer is up to date and all decrees have been written into the
ledger.  Therefore we should allow the proposer to retry by reevaluating
the decree number and resending a proposal.