Fix case of bootstrapping highest proposed decree content
Issue is that if bootstrapping a replica includes a highest proposed
decree with non-empty content then subsequent proposals will stall while
waiting for the non-empty decree to be promised and accepted. In
actuality the boostrapped replica should have no business waiting to
flush out this non-empty decree because it must have originated from the
old replica. Instead we should let the old replica handle that decree.