Narrow the scope of returning promise based on previous accepted value
Issue is that if we receive a prepare for lower decree than previous
highest promised decree then we should not send promise even if there is
already a pending accept decree. There is no point to send promise with
a previously accepted decree as we know that the prepared decree is too
low and will never be accepted. Instead we should nack and wait for
proposer to increment and send higher prepare message.